pyproject.toml
[tool.poetry]
name = "tech-trends-chatbot-backend"
version = "0.1.0"
description = ""
authors = ["ruizguille <guillermo@codeawake.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
pdfminer-six = "^20240706"
tqdm = "^4.66.4"
tiktoken = "^0.7.0"
nltk = "^3.8.1"
pydantic-settings = "^2.3.4"
redis = "^5.0.8"
numpy = "^2.0.1"
pydantic = "^2.8.2"
rich = "^13.7.1"
openai = "^1.40.1"
fastapi = {extras = ["standard"], version = "^0.112.0"}

sse-starlette = "^2.1.3"
google-generativeai = "^0.8.4"
poetry-plugin-export = "^1.9.0"
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
load = "app.loader:main"
local = "app.assistants.local_assistant:main"
export = "app.export:main"